Table of Contents final
uses 3 different sets of arrays within an array. (correct phrasing?)

for each object in the TABLE array.. point the variable row.
FIRST gives us the left half of the page, String 'Chapter ' then adds a
chapternumber using chapstart.to_s and pulls a chapter name from our
array.. it chooses the 0th variable from each of the 3 arrays.

LAST gives us the string "Page '  and adds a page number using
pagenum.to_s .. this is pulling the "1st" aka the second variable in
each of the 3 arrays.

finally the puts command aligns what we've done and chapstart =
chapstart + 1 tells us that we should be adding 1 everytime we pull a
chapstart..